// types/home.ts
export interface NowPlayingParams {
  language?: string;
  api_key?: string;
  page?: number;
}
// ...
export interface MovieDetailParams {
  movie_id: number;
  language?: string;
  api_key?: string;
}

export interface CharacterDetailParams {
  person_id: number;
  language?: string;
  api_key?: string;
}